fix(rust): clear the 12 clippy warnings
Nothing here was a defect, but the lints have to be at zero before CI
can enforce them.

Most were redundant u8 casts and needless borrows. Two were worth more
than the diff suggests: the Arrow capsule name was a byte string with a
hand-written nul terminator, now a c"" literal that cannot lose it, and
the year bounds check became a range contains.

Dropping the casts also means a future pyo3 return-type change becomes a
compile error rather than a silent truncation.

+++ src/helpers.rs
@@ -27,8 +27,8 @@ pub enum ColType {
 pub fn py_datetime_to_excel(dt: &Bound<PyDateTime>) -> PyResult<ExcelDateTime> {
     ExcelDateTime::from_ymd(
         dt.get_year() as u16,
-        dt.get_month() as u8,
-        dt.get_day() as u8,
+        dt.get_month(),
+        dt.get_day(),
     )
     .map_err(|e| {
         PyErr::new::<pyo3::exceptions::PyValueError, _>(format!(
@@ -38,8 +38,8 @@ pub fn py_datetime_to_excel(dt: &Bound<PyDateTime>) -> PyResult<ExcelDateTime> {
     })?
     .and_hms(
         dt.get_hour() as u16,
-        dt.get_minute() as u8,
-        dt.get_second() as u8,
+        dt.get_minute(),
+        dt.get_second(),
     )
     .map_err(|e| {
         PyErr::new::<pyo3::exceptions::PyValueError, _>(format!(
@@ -53,8 +53,8 @@ pub fn py_datetime_to_excel(dt: &Bound<PyDateTime>) -> PyResult<ExcelDateTime> {
 pub fn py_date_to_excel(d: &Bound<PyDate>) -> PyResult<ExcelDateTime> {
     ExcelDateTime::from_ymd(
         d.get_year() as u16,
-        d.get_month() as u8,
-        d.get_day() as u8,
+        d.get_month(),
+        d.get_day(),
     )
     .map_err(|e| {
         PyErr::new::<pyo3::exceptions::PyValueError, _>(format!(
